refactor(evaluate/io): move private _*.py modules into io/data/ subpackage
